STYLE: vtkMRMLScalarVolumeDisplayNode: Fix "narrowing conversion" warning
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=26002
git-svn-id: http://svn.slicer.org/Slicer4/trunk@26002 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkMRMLScalarVolumeDisplayNode.cxx
@@ -799,7 +799,7 @@ void vtkMRMLScalarVolumeDisplayNode::CalculateAutoLevels()
     this->Accumulate->SetInputData(imageDataScalar);
     int extent[6] = {0, 999, 0, 0, 0, 0};
     this->Accumulate->SetComponentExtent(extent);
-    double origin[3] = {minInt, 0, 0};
+    double origin[3] = {static_cast<double>(minInt), 0.0, 0.0};
     this->Accumulate->SetComponentOrigin(origin);
     double spacing[3] = {(maxInt-minInt)/1000.0, 1.0, 1.0};
     this->Accumulate->SetComponentSpacing(spacing);
